requirements.txt
# streamlit==0.1.0
streamlit==1.30.0
requests==2.31.
transformers==4.36.2
python-dotenv==1.0.1
torch==2.1.2
ipython==8.18.1
pandas==2.1.4